Refactor: Create player module structure and move Player type
