+++ src/websocket_server/modules/player/player.types.ts
@@ -1,0 +1,5 @@
+export interface Player {
+    id: string;
+    name: string;
+    password: string;
+}
